refactor(admin): migrate hashtags tab to canonical /social/... path
